Renamed `QueryAnswers` to `QueryQuestionAnswer`
https://coveord.atlassian.net/browse/JSUI-2949
